add maxudpsize option
allows to control the maximum udp packet size. when pipelining a lot
of metrics within a controlled network or to 127.0.0.1 it can be desirable
to lower the number of udp packets by sending larger onces.

the documentation reflects that this option is not for the faint hearted,
should not be used on the internet and should be handled with extreme care.