Ignore RuntimeErrors together with socket errors
When Python's standard library is patched with Eventlet the exception
raised when two threads access the same socket simultaneously is
RuntimeError and I think it makes sense to include it here.

The downside is it may start ignoring too wide variety of exceptions but
on the other hand I can't think of any other RuntimeErrors that can be
raised here.