Add helper to instantiate from environment
In statsd 2.0 era, you used to "import statsd" which would set the
global statsd object to either a client, or, if no environment
variables were detected, to "None".  In 3.0 this moved into
"statsd.defaults.env" (where the semantics changed to use default
values).

So we have a bunch of code that relies on getting the statsd client
populated from environment variables from this legacy.

However, setting a global statsd object at import time has some
problems; for example it interferes with unit-testing where we want to
be able to call individual classes with different statsd targets to
make sure we're capturing their stats output for correctness.

So we can move our objects to creating their own StatsClient objects,
but there's no interface for instantiating that with values from the
environment.  Thus we end up writing a wrapper which has been
duplicated a few times.

This proposes such a wrapper for statsd directly.  With it, one can do

 def __init__():
   ...
   self.statsd = statsd.defaults.from_env()

in their object and "self.statsd" will follow the semantics of
"statsd.defaults.env" for creating a StatsClient object.